requirements_DisCo.txt
absl-py==1.0.0
addict==2.2.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
alabaster==0.7.16
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
args==0.1.0
astor==0.8.1
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
asynctest==0.13.0
attrs==21.4.0
autograd==1.3
autograd-gamma==0.5.0
Automat==22.10.0
autopep8==1.6.0
babel==2.17.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==4.1.0
blessed==1.20.0
buildtools==1.0.6
cached-property==1.5.2
cachetools==4.2.4
certifi==2024.2.2
cffi==1.14.6
charset-normalizer==2.0.9
click==8.0.4
clint==0.5.1
colorama==0.4.6
comet-ml==3.40.0
comm==0.2.2
configobj==5.0.8
constantly==23.10.4
contextlib2==21.6.0
coverage==7.4.4
cycler==0.11.0
Cython==3.0.8
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
DingtalkChatbot==1.5.7
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.21.2
duckdb==1.2.1
dulwich==0.21.7
ecos==2.0.12
einops==0.4.1
einops-exts==0.0.4
entrypoints==0.3
et-xmlfile==1.1.0
everett==3.1.0
except==1.0a1
exceptiongroup==1.2.0
executing==2.0.1
faiss-gpu==1.7.2
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.3
formulaic==0.2.4
frozenlist==1.3.3
fsspec==2024.3.1
furl==2.1.3
future==0.18.2
geopandas==1.0.1
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.3.3
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
gpustat==1.1.1
gradio==4.26.0
gradio_client==0.15.1
greenlet==3.0.3
grpcio==1.43.0
h11==0.14.0
h5py==3.1.0
histocartography==0.1.4
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
hyperlink==21.0.0
idc-index==0.8.6
idc-index-data==20.0.3
idna==3.3
imageio==2.15.0
imagesize==1.4.1
importlib_metadata==8.6.1
importlib_resources==6.4.0
incremental==22.10.0
interface-meta==1.2.4
intervaltree==3.0.2
ipykernel==5.5.6
ipython==8.12.3
ipython-genutils==0.2.0
isodate==0.6.1
jedi==0.17.2
Jinja2==3.1.6
joblib==1.1.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
jupyter-client==7.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.9.1
jupyter-highlight-selected-word==0.2.0
jupyter_latex_envs==1.4.6
jupyter_nbextensions_configurator==0.4.1
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
lief==0.11.5
lifelines==0.26.4
loguru==0.7.2
lxml==4.7.1
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.3.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml_collections==0.1.1
multidict==6.0.4
nbclient==0.5.9
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.5.4
networkx==2.5.1
ninja==1.11.1.1
nmslib==2.1.1
notebook==6.4.6
numexpr==2.8.6
numpy==1.23.0
nvidia-ml-py==12.535.133
nvitop==1.3.2
nystrom-attention==0.0.11
oauthlib==3.1.1
olefile==0.47
opencv-python==4.5.5.62
opencv-python-headless==4.5.5.62
openpyxl==3.1.2
openslide-python==1.1.2
opt-einsum==3.3.0
orderedmultidict==1.0.1
orjson==3.10.0
osqp==0.6.3
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pip==24.2
pipreqs==0.5.0
platformdirs==4.2.0
pretty-errors==1.2.25
progressbar2==4.5.0
prometheus-client==0.12.0
prompt-toolkit==3.0.43
protobuf==3.19.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.1
pycodestyle==2.8.0
pycparser==2.21
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyogrio==0.10.0
pyparsing==3.0.6
pyproj==3.6.1
pyproject-toml==0.0.10
pyrsistent==0.18.0
python-box==6.1.0
python-dateutil==2.8.2
python-multipart==0.0.9
python-utils==3.9.1
pytorch-lightning==1.2.3
pytorch_toolbelt==0.4.0
pytz==2021.3
pyvips==2.1.15
PyWavelets==1.1.1
PyYAML==6.0.2
pyzmq==22.3.0
qdldl==0.1.7.post0
rdflib==5.0.0
redis==5.0.3
redo==2.0.4
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.8
ruff==0.3.7
s5cmd==0.2.0
safetensors==0.4.2
SciencePlots==2.1.1
scikit-learn==1.0.2
scikit-survival==0.17.2
scipy==1.5.4
seaborn==0.11.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==1.20.0
setproctitle==1.2.3
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.4.7
sphinx-click==6.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.29
stack-data==0.6.3
staintools==2.1.2
starlette==0.37.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.0.0
tifffile==2020.9.3
timm==1.0.3
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1+cu116
torch-cluster==1.6.1+pt113cu116
torch_geometric==2.5.2
torch-scatter==2.1.1+pt113cu116
torch-sparse==0.6.17+pt113cu116
torch-spline-conv==1.2.2+pt113cu116
torchnet==0.0.4
torchstain==1.3.0
torchsummary==1.5.1
torchvision==0.14.1+cu116
tornado==6.1
tqdm==4.62.3
traitlets==5.14.2
transformers==4.31.0
triton==2.2.0
Twisted==24.3.0
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.2
urllib3==1.26.15
uvicorn==0.29.0
visdom==0.2.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==11.0.3
Werkzeug==2.0.2
wheel==0.44.0
wrapt==1.13.3
wurlitzer==3.0.3
yacs==0.1.8
yarg==0.1.9
yarl==1.9.4
zipp==3.21.0
zope.interface==6.2
